requirements.txt
ansys-api-mapdl==0.5.2
ansys-api-platform-instancemanagement==1.1.3
ansys-mapdl-core==0.70.2
ansys-mapdl-reader==0.55.1
ansys-math-core==0.2.4
ansys-platform-instancemanagement==1.1.2
ansys-tools-path==0.7.3
appdirs==1.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
geomdl==5.4.0
grpcio==1.73.1
idna==3.10
importlib_metadata==8.7.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pyansys-tools-versioning==0.6.0
pyiges==0.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvista==0.45.2
requests==2.32.4
scipy==1.16.0
scooby==0.10.1
six==1.17.0
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vtk==9.4.2
zipp==3.23.0
